Refactor : Directory 구조 변경 entity + repository -> domain
