+++ server/DreamConsumer-main/src/main/java/com/example/demo/user/domain/UserRepository.java
@@ -1,0 +1,10 @@
+package com.example.demo.user.repository;
+
+import com.example.demo.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
